requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv==0.13.0
pymysql==0.9.3